// features/admin.tenants.v1/models/system-settings/admin-advisory.ts
/**
 * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
 *
 * WSO2 LLC. licenses this file to you under the Apache License,
 * Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/**
 * Interface for the admin advisory banner configuration.
 */
export interface AdminAdvisoryBannerConfigurationInterface {
    /**
     * Text content of the banner.
     */
    bannerContent: string;
    /**
     * Is the banner enabled or not.
     */
    enableBanner: boolean;
}